+++ firebase/user.py
@@ -1,5 +1,4 @@
 from firebase.setup import db, auth
-from firebase_admin import firestore
 
 
 def register(email, password, display_name):