+++ static/js/checkScroll.js
@@ -2,17 +2,16 @@ function headerState() {
     if (window.scrollY > 100) {
         document.getElementById("header").style.boxShadow = "var(--bs)";
         document.getElementById("header").style.background = "var(--bg-alt)";
-    }
-    else {
+    } else {
         document.getElementById("header").style.boxShadow = "unset";
         document.getElementById("header").style.background = "var(--bg)";
     }
 }
 
-document.addEventListener('scroll', function(e) {
+document.addEventListener('scroll', function (e) {
     lastKnownScrollPosition = window.scrollY;
-        headerState()
-    });
+    headerState()
+});
 
 (function () {
     headerState();