+++ static/js/disableOnURL.js
@@ -1,8 +1,8 @@
-(function() {
+(function () {
     var parts = window.location.href.split('/');
     var lastSegment = parts.pop() || parts.pop();
     console.log(lastSegment);
-    if (lastSegment = 'about'){
+    if (lastSegment = 'about') {
         console.log(document.getElementById('about-header'))
         // document.getElementById('about-header').style.pointerEvents = 'none';
     }