+++ static/js/theme.js
@@ -2,19 +2,19 @@ function setTheme(themeName) {
     localStorage.setItem('theme', themeName);
     document.documentElement.className = themeName;
 }
+
 function toggleTheme() {
     if (localStorage.getItem('theme') === 'theme-light') {
         setTheme('theme-dark');
-    }
-    else {
+    } else {
         setTheme('theme-light');
     }
 }
+
 (function () {
     if (localStorage.getItem('theme') === 'theme-dark') {
         setTheme('theme-dark');
-    }
-    else {
+    } else {
         setTheme('theme-light');
     }
 })();