+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - implement login and register in seperate urls as well as index.html popups
 - implement profile picture backend
 - auto-redirect http to https for security reasons
+- Security Updates
+- performance improvements
 
 ### 0.0.3-dev
 - Organize CSS