+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - implement article backend
 - register and log in has seperate urls
 - performance improvements
+- custom scrollbars
 
 ### 0.0.4-dev
 - implement login and register in seperate urls as well as index.html popups