+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 ## Changelog
 
 ### 0.0.5-dev-pre
+
 - implement article backend
 - register and log in has seperate urls
 - performance improvements
 - custom scrollbars
+- code reformatting
 
 ### 0.0.4-dev
+
 - implement login and register in seperate urls as well as index.html popups
 - implement profile picture backend
 - auto-redirect http to https for security reasons
@@ -14,15 +17,18 @@
 - performance improvements
 
 ### 0.0.3-dev
+
 - Organize CSS
 - Remove test code
 - Visual changes including animations
 - add verification frontend to register page
 
 ### 0.0.2-dev
+
 - Implement Continous Integration
 - Update Dependencies
 
 ### 0.0.1-dev
+
 - First dev release.
 - Merged initial backend and ui